requirements.txt
mcp>=1.9.0
litellm>=1.0.0
python-dotenv>=1.0.0
